Updated also the do while is back in..
So I reworked the code and I added back the do while.. I didn't save the previous one so I cant remember if I changed much with it. As far as I can tell it works I've tested it multiple times and it starts back where its supposed to. I even stepped through an the compiler was happy.